Split roboOutputName into another file
