+++ include-build/roborazzi-core/src/commonJvmMain/kotlin/com/github/takahirom/roborazzi/DefaultFileNameGenerator.kt
@@ -119,17 +119,3 @@ object DefaultFileNameGenerator {
     }
   }
 }
-
-/**
- * You can utilize this function to generate the file name of the image to be recorded.
- * You can change naming strategy by setting roborazzi.record.namingStrategy.
- */
-@ExperimentalRoborazziApi
-fun roboOutputName(): String {
-  val roborazziContext = provideRoborazziContext()
-  val description = roborazziContext.description
-  if (description != null) {
-    return DefaultFileNameGenerator.generateOutputNameWithDescription(description)
-  }
-  return DefaultFileNameGenerator.generateOutputNameWithStackTrace()
-}
